Move progressbar display back to the editor window
The rename asset operation was where the real slowdown occurred,
so I moved the display back to the window.
